Tests/LibWeb/Text/expected/css/computed-style-without-layout-box.txt

contents display: contents
contents color: rgb(10, 20, 30)
contents line-height: 1.5
contents margins: 1px 2px 9px 4px
contents margin shorthand: 1px 2px 9px 4px
contents logical margin: 9px
contents background image: linear-gradient(currentcolor, rgb(40, 50, 60))
none display: none
none color: rgb(70, 80, 90)
none line-height: normal
none padding shorthand: 5px 6px 7px 8px
pseudo display: none
pseudo color: rgb(100, 110, 120)
pseudo margin shorthand: 11px 12px 13px 14px
